chore: storybook fixes
