+++ src/components/Form/FormInputGroup.examples.md
@@ -3,7 +3,7 @@ import { Button, Form } from "tabler-react";
 
 <Form.InputGroup
   append={
-    <Button RootComponent="a" color="primary" href="http://www.google.com">
+    <Button as="a" color="primary" href="http://www.google.com">
       Go!
     </Button>
   }
@@ -18,7 +18,7 @@ import { Button, Form } from "tabler-react";
 <Form.InputGroup
   inputProps={{ placeholder: "Search for..." }}
   append={
-    <Button RootComponent="a" color="primary" href="http://www.google.com">
+    <Button as="a" color="primary" href="http://www.google.com">
       Go!
     </Button>
   }
