+++ src/components/Form/FormInputGroupAppend.examples.md
@@ -4,7 +4,7 @@ import { Button, Form } from "tabler-react";
 <Form.InputGroup>
   <Form.Input placeholder="Search for..." />
   <Form.InputGroupAppend>
-    <Button RootComponent="a" color="primary" href="http://www.google.com">
+    <Button as="a" color="primary" href="http://www.google.com">
       Go!
     </Button>
   </Form.InputGroupAppend>
