+++ src/components/Form/FormInputGroupAppend.stories.tsx
@@ -18,11 +18,7 @@ stories.add(
       <Form.InputGroup>
         <Form.Input placeholder="Search for..." />
         <Form.InputGroupAppend>
-          <Button
-            RootComponent="a"
-            color="primary"
-            href="http://www.google.com"
-          >
+          <Button as="a" color="primary" href="http://www.google.com">
             Go!
           </Button>
         </Form.InputGroupAppend>
