+++ src/components/Form/FormInputGroupPrepend.stories.tsx
@@ -18,11 +18,7 @@ stories.add(
       <Form.InputGroupPrepend>
         <Form.Input placeholder="Search for..." />
         <Form.InputGroupPrepend>
-          <Button
-            RootComponent="a"
-            color="primary"
-            href="http://www.google.com"
-          >
+          <Button as="a" color="primary" href="http://www.google.com">
             Go!
           </Button>
         </Form.InputGroupPrepend>
