+++ src/components/Grid/GridCol.stories.tsx
@@ -38,9 +38,9 @@ stories.add(
 stories.add("Column Width", () => {
   return (
     <Grid.Row>
-      <Grid.Col width={2}>1</Grid.Col>
-      <Grid.Col width={4}>2</Grid.Col>
-      <Grid.Col width={6}>3</Grid.Col>
+      <Grid.Col col={2}>1</Grid.Col>
+      <Grid.Col col={4}>2</Grid.Col>
+      <Grid.Col col={6}>3</Grid.Col>
     </Grid.Row>
   );
 });
@@ -48,22 +48,22 @@ stories.add("Column Width", () => {
 stories.add("Responsive column widths", () => {
   return (
     <Grid.Row>
-      <Grid.Col xl={1} lg={2} md={4} sm={6} xs={12}>
+      <Grid.Col xxl={1} xl={2} lg={4} md={6} sm={12}>
         1
       </Grid.Col>
-      <Grid.Col xl={1} lg={2} md={4} sm={6} xs={12}>
+      <Grid.Col xxl={1} xl={2} lg={4} md={6} sm={12}>
         2
       </Grid.Col>
-      <Grid.Col xl={1} lg={2} md={4} sm={6} xs={12}>
+      <Grid.Col xxl={1} xl={2} lg={4} md={6} sm={12}>
         3
       </Grid.Col>
-      <Grid.Col xl={1} lg={2} md={4} sm={6} xs={12}>
+      <Grid.Col xxl={1} xl={2} lg={4} md={6} sm={12}>
         4
       </Grid.Col>
-      <Grid.Col xl={1} lg={2} md={4} sm={6} xs={12}>
+      <Grid.Col xxl={1} xl={2} lg={4} md={6} sm={12}>
         5
       </Grid.Col>
-      <Grid.Col xl={1} lg={2} md={4} sm={6} xs={12}>
+      <Grid.Col xxl={1} xl={2} lg={4} md={6} sm={12}>
         6
       </Grid.Col>
     </Grid.Row>
@@ -73,7 +73,7 @@ stories.add("Responsive column widths", () => {
 stories.add("Column Offset", () => {
   return (
     <Grid.Row>
-      <Grid.Col width={6} offset={3}>
+      <Grid.Col col={6} offset={3}>
         2
       </Grid.Col>
       <Grid.Col md={6} offsetMd={3}>
